pyproject.toml
[project]
name = "vibe-marketing"
version = "0.1.0"
description = "Instagram Reel → Business Strategy Pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "anthropic>=0.42.0",
    "faster-whisper>=1.1.0",
    "httpx>=0.28.0",
    "loguru>=0.7.0",
    "yt-dlp>=2024.12.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
